+++ tests/testthat/test-ggplot-bar.R
@@ -2,7 +2,7 @@ context("bar")
 
 expect_traces <- function(gg, n.traces, name) {
   stopifnot(is.numeric(n.traces))
-  L <- save_outputs(gg, paste0("bar-", name))
+  L <- expect_doppelganger_built(gg, paste0("bar-", name))
   all.traces <- L$data
   no.data <- sapply(all.traces, function(tr) {
     is.null(tr[["x"]]) && is.null(tr[["y"]])